In frontend, move legacy/* at the root
