+++ packages/frontend/algolia.js
