+++ packages/frontend/feedSerializer.js
